requirements.txt
python-dotenv
ccxt
backtrader
pandas
numpy
scikit-learn
torch
python-telegram-bot
river
